handle xml readable interfaces for binary trajectory
