+++ src/libopenrave-core/generictrajectory.cpp
@@ -23,7 +23,7 @@ namespace OpenRAVE {
 
 // To distinguish between binary and XML trajectory files
 static const uint16_t MAGIC_NUMBER = 0x62ff;
-static const uint16_t VERSION_NUMBER = 0x0002;  // Version number for serialization
+static const uint16_t BINARY_TRAJECTORY_VERSION_NUMBER = 0x0003;  // Version number for serialization
 
 static const dReal g_fEpsilonLinear = RavePow(g_fEpsilon,0.9);
 static const dReal g_fEpsilonQuadratic = RavePow(g_fEpsilon,0.45); // should be 0.6...perhaps this is related to parabolic smoother epsilons?
@@ -421,7 +421,7 @@ public:
 
             // Write binary file header
             WriteBinaryUInt16(O, MAGIC_NUMBER);
-            WriteBinaryUInt16(O, VERSION_NUMBER);
+            WriteBinaryUInt16(O, BINARY_TRAJECTORY_VERSION_NUMBER);
 
             /* Store meta-data */
 
@@ -443,19 +443,26 @@ public:
 
             WriteBinaryString(O, GetDescription());
 
-            // Readable interfaces, added on VERSION_NUMBER=0x0002
+            // Readable interfaces, added on BINARY_TRAJECTORY_VERSION_NUMBER=0x0002
             std::stringstream ss;
             const uint16_t numReadableInterfaces = GetReadableInterfaces().size();
             WriteBinaryUInt16(O, numReadableInterfaces);
             FOREACHC(itReadableInterface, GetReadableInterfaces()) {
                 WriteBinaryString(O, itReadableInterface->first);  // xmlid
 
-                xmlreaders::StreamXMLWriterPtr writer(new xmlreaders::StreamXMLWriter(std::string()));
-                itReadableInterface->second->Serialize(writer, options);
-                ss.clear();
+                // try to see if the readable interface is a known type                ss.clear();
                 ss.str(std::string());
+                xmlreaders::StreamXMLWriterPtr writer(new xmlreaders::StreamXMLWriter(itReadableInterface->second->GetXMLId()));
+                itReadableInterface->second->Serialize(writer, options);
                 writer->Serialize(ss);
+
+                std::string readerType;
+                xmlreaders::HierarchicalXMLReadablePtr pHierarchical = boost::dynamic_pointer_cast<xmlreaders::HierarchicalXMLReadable>(itReadableInterface->second);
+                if( !!pHierarchical ) {
+                    readerType = "HierarchicalXMLReadable";
+                }
                 WriteBinaryString(O, ss.str());
+                WriteBinaryString(O, readerType);
             }
         }
     }
@@ -476,7 +483,7 @@ public:
             ReadBinaryUInt16(I, versionNumber);
 
             // currently supported versions: 0x0001, 0x0002
-            if (versionNumber > VERSION_NUMBER || versionNumber < 0x0001)
+            if (versionNumber > BINARY_TRAJECTORY_VERSION_NUMBER || versionNumber < 0x0001)
             {
                 throw OPENRAVE_EXCEPTION_FORMAT(_("unsupported trajectory format version %d "),versionNumber,ORE_InvalidArguments);
             }
@@ -515,14 +522,28 @@ public:
                 // read readable interfaces
                 uint16_t numReadableInterfaces = 0;
                 ReadBinaryUInt16(I, numReadableInterfaces);
-                std::string xmlid;
+                std::string xmlid, readerType;
                 std::string serializedReadableInterface;
                 for (size_t readableInterfaceIndex = 0; readableInterfaceIndex < numReadableInterfaces; ++readableInterfaceIndex)
                 {
                     ReadBinaryString(I, xmlid);
                     ReadBinaryString(I, serializedReadableInterface);
 
-                    XMLReadablePtr readableInterface(new xmlreaders::StringXMLReadable(xmlid, serializedReadableInterface));
+                    XMLReadablePtr readableInterface;
+                    if( versionNumber >= 3 ) {
+                        ReadBinaryString(I, readerType);
+                        if( readerType == "HierarchicalXMLReadable" ) {     
+                            xmlreaders::HierarchicalXMLReader xmlreader(xmlid, AttributesList());
+                            xmlreaders::ParseXMLData(xmlreader, serializedReadableInterface.c_str(), serializedReadableInterface.size());
+                            readableInterface = xmlreader.GetReadable();
+                        }
+                        else {
+                            readableInterface.reset(new xmlreaders::StringXMLReadable(xmlid, serializedReadableInterface));
+                        }
+                    }
+                    else {
+                        readableInterface.reset(new xmlreaders::StringXMLReadable(xmlid, serializedReadableInterface));
+                    }
                     SetReadableInterface(xmlid, readableInterface);
                 }
             }
